# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Amazing Studio - Wedding Studio Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: assigneeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    put:
      operationId: updateQuote
      tags: [quotes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /services:
    get:
      operationId: listServices
      tags: [services]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    put:
      operationId: updateStaff
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        totalBookings:
          type: integer
        totalDebt:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, createdAt, totalBookings, totalDebt]
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, phone]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        assigneeId:
          type: integer
          nullable: true
        assigneeName:
          type: string
          nullable: true
        bookingId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        dueDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, priority, status, createdAt]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        assigneeId:
          type: integer
          nullable: true
        bookingId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date
          nullable: true
      required: [title, priority]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        assigneeId:
          type: integer
          nullable: true
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [todo, in_progress, done, cancelled]
        dueDate:
          type: string
          format: date
          nullable: true
    Quote:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        totalAmount:
          type: number
        discount:
          type: number
        finalAmount:
          type: number
        status:
          type: string
          enum: [draft, sent, accepted, rejected]
        validUntil:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, customerId, customerName, customerPhone, title, items, totalAmount, discount, finalAmount, status, createdAt]
    QuoteItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
      required: [name, quantity, unitPrice, total]
    CreateQuoteRequest:
      type: object
      properties:
        customerId:
          type: integer
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        discount:
          type: number
        validUntil:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required: [customerId, title, items, discount]
    UpdateQuoteRequest:
      type: object
      properties:
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        discount:
          type: number
        status:
          type: string
          enum: [draft, sent, accepted, rejected]
        validUntil:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [package, addon]
        price:
          type: number
        duration:
          type: string
          nullable: true
        includes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, price, includes, isActive, createdAt]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [package, addon]
        price:
          type: number
        duration:
          type: string
          nullable: true
        includes:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [name, type, price, includes, isActive]
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, photographer, assistant, receptionist]
        email:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        joinDate:
          type: string
          format: date
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, role, joinDate, isActive, createdAt]
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, photographer, assistant, receptionist]
        email:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        joinDate:
          type: string
          format: date
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
      required: [name, phone, role, joinDate, isActive]
...