# lib/api-zod/src/generated/types/createQuoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazing Studio - Wedding Studio Management API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItem } from "./quoteItem";

export interface CreateQuoteRequest {
  customerId: number;
  title: string;
  items: QuoteItem[];
  discount: number;
  validUntil?: Date | null;
  notes?: string | null;
}