# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazing Studio - Wedding Studio Management API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItem } from "./quoteItem";
import type { QuoteStatus } from "./quoteStatus";

export interface Quote {
  id: number;
  customerId: number;
  customerName: string;
  customerPhone: string;
  title: string;
  items: QuoteItem[];
  totalAmount: number;
  discount: number;
  finalAmount: number;
  status: QuoteStatus;
  validUntil?: Date | null;
  notes?: string | null;
  createdAt: Date;
}