# lib/api-zod/src/generated/types/updateQuoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amazing Studio - Wedding Studio Management API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteItem } from "./quoteItem";
import type { UpdateQuoteRequestStatus } from "./updateQuoteRequestStatus";

export interface UpdateQuoteRequest {
  title?: string;
  items?: QuoteItem[];
  discount?: number;
  status?: UpdateQuoteRequestStatus;
  validUntil?: Date | null;
  notes?: string | null;
}